Move frontend scripts to src
