+++ index.html
@@ -41,7 +41,7 @@
 
     <!-- Preload Critical Assets -->
     <link rel="preload" href="assets/images/web/logo.webp" as="image" type="image/webp" fetchpriority="high">
-    <link rel="preload" href="assets/js/script.js" as="script">
+      <link rel="preload" href="assets/js/script.min.js" as="script">
 
     <!-- Deferred Non-Critical CSS -->
     <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.3/dist/css/bootstrap.min.css" media="print" data-defer integrity="sha384-QWTKZyjpPEjISv5WaRU9OFeRpok6YctnYmDr5pNlyT2bRjXh0JMhjY6hW+ALEwIH" crossorigin="anonymous">
@@ -137,7 +137,7 @@
     <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.3/dist/js/bootstrap.min.js"
         integrity="sha384-0pUGZvbkm6XF6gxjEnlmuGrJXVbNuzT9qBBavbLwCsOGabYfZo0T0to5eqruptLy"
         crossorigin="anonymous"></script>
-    <script src="assets/js/script.js" defer></script>
+    <script src="assets/js/script.min.js" defer></script>
 
     <!-- Service Worker registration handled in assets/js/script.min.js -->
 </body>
