+++ src/js/modules/perf.js
